+++ source/light_shading/normal_at.c
@@ -6,7 +6,7 @@
 /*   By: rcatini <rcatini@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/21 12:12:37 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/03/24 15:58:19 by rcatini          ###   ########.fr       */
+/*   Updated: 2023/03/25 16:48:25 by rcatini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ t_tuple	normal_at(t_object object, t_tuple world_point)
 		object_normal = (t_tuple){{0, 1, 0, 0}};
 	else
 	{
-		dist = powf(object_point.x, 2) + powf(object_point.z, 2);
+		dist = pow(object_point.x, 2) + pow(object_point.z, 2);
 		if (dist < 1 && object_point.y >= 1 - EPSILON)
 			object_normal = (t_tuple){{0, 1, 0, 0}};
 		else if (dist < 1 && object_point.y <= -1 + EPSILON)
