+++ source/transformation/rotation_xyz.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   rotation_xyz.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tda-silv <tda-silv@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rcatini <rcatini@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/20 01:13:56 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/03/24 08:50:02 by tda-silv         ###   ########.fr       */
+/*   Updated: 2023/03/25 16:47:05 by rcatini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@ t_matrix4	rotation_x(double deg)
 
 	rad = deg_to_rad(deg);
 	mtx = give_identity_matrix4();
-	mtx.tab[1][1] = cosf(rad);
-	mtx.tab[1][2] = sinf(rad) * -1;
-	mtx.tab[2][1] = sinf(rad);
-	mtx.tab[2][2] = cosf(rad);
+	mtx.tab[1][1] = cos(rad);
+	mtx.tab[1][2] = sin(rad) * -1;
+	mtx.tab[2][1] = sin(rad);
+	mtx.tab[2][2] = cos(rad);
 	return (mtx);
 }
 
@@ -33,10 +33,10 @@ t_matrix4	rotation_y(double deg)
 
 	rad = deg_to_rad(deg);
 	mtx = give_identity_matrix4();
-	mtx.tab[0][0] = cosf(rad);
-	mtx.tab[0][2] = sinf(rad);
-	mtx.tab[2][0] = sinf(rad) * -1;
-	mtx.tab[2][2] = cosf(rad);
+	mtx.tab[0][0] = cos(rad);
+	mtx.tab[0][2] = sin(rad);
+	mtx.tab[2][0] = sin(rad) * -1;
+	mtx.tab[2][2] = cos(rad);
 	return (mtx);
 }
 
@@ -47,9 +47,9 @@ t_matrix4	rotation_z(double deg)
 
 	rad = deg_to_rad(deg);
 	mtx = give_identity_matrix4();
-	mtx.tab[0][0] = cosf(rad);
-	mtx.tab[0][1] = sinf(rad) * -1;
-	mtx.tab[1][0] = sinf(rad);
-	mtx.tab[1][1] = cosf(rad);
+	mtx.tab[0][0] = cos(rad);
+	mtx.tab[0][1] = sin(rad) * -1;
+	mtx.tab[1][0] = sin(rad);
+	mtx.tab[1][1] = cos(rad);
 	return (mtx);
 }
