+++ source/tuple/normalization_vector.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   normalization_vector.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tda-silv <tda-silv@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rcatini <rcatini@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/18 17:56:06 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/03/04 19:09:48 by tda-silv         ###   ########.fr       */
+/*   Updated: 2023/03/25 16:49:46 by rcatini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ t_tuple	normalization_vector(t_tuple tuple)
 	t_tuple	ret;
 	double	len;
 
-	len = sqrtf(tuple.x * tuple.x + tuple.y * tuple.y + tuple.z * tuple.z);
+	len = sqrt(tuple.x * tuple.x + tuple.y * tuple.y + tuple.z * tuple.z);
 	ret.x = tuple.x / len;
 	ret.y = tuple.y / len;
 	ret.z = tuple.z / len;
